networking kind of works
client cannot properly receive messages... probably have to listen on a different port while server connects to it.
column limits were added and are good
attempting to have server add to columns will be a bit hard since most of that is handled in .kv file. but if you just search for when StringProperty changes in those columns it should be fine.